fix(indexer): Rename types.rs to client.rs
